+++ src/categories/categories.controller.ts
@@ -2,7 +2,7 @@ import { Controller, Get, Injectable, Query, UseFilters } from '@nestjs/common';
 import { Sequelize } from 'sequelize-typescript';
 import { ApiQuery, ApiResponse, ApiTags } from '@nestjs/swagger';
 
-import { ErrorHandler } from '../error/errorHandler';
+import { ErrorHandler } from '../error/error-handler';
 import { CategoriesService } from './services/categories.service';
 import { Category } from './models/categories.model';
 import { Language } from '../types/language.type.enum';
