+++ src/error/forbidden-operation.error.ts
@@ -1,0 +1,5 @@
+export default class ForbiddenOperationError extends Error {
+  constructor(message = 'FORBIDDEN_OPERATION') {
+    super(message);
+  }
+}
