+++ src/places/services/places.service.ts
@@ -3,8 +3,8 @@ import { InjectModel } from '@nestjs/sequelize';
 import { Transaction } from 'sequelize';
 
 import { Place } from '../models/places.model';
-import { CreatePlaceDto } from '../dto/createPlaceDto';
-import { UpdatePlaceDto } from '../dto/updatePlaceDto';
+import { CreatePlaceDto } from '../dto/create-place.dto';
+import { UpdatePlaceDto } from '../dto/update-place.dto';
 import { UsersService } from '../../users/users.service';
 import CRUDError from '../../error/CRUD.error';
